fix(fe): fix strange dynamic/static rendering behaviour
